% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimIA-package.R
\docType{package}
\name{optimIA-package}
\alias{optimIA-package}
\alias{optimIA}
\title{The 'optimIA' package.}
\description{
The R Package for Optimizing Interim Analysis Timing for Bayesian Adaptive Commensurate Designs
}
\references{
Wu, X., Xu, Y., & Carlin, B. P. (2019). Optimizing Interim Analysis Timing for Bayesian Adaptive Commensurate Designs. arXiv preprint arXiv:1905.07456.
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
