% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCDOT_function.R
\name{optimIA}
\alias{optimIA}
\title{Optimizing Interim Analysis Timing for Bayesian Adaptive Commensurate Designs}
\usage{
optimIA(historical, w = 0.5, IA_num, N, p_l, p_0, p_u, delta_0, delta_1,
  delta_design, delta_min, type1_level = 0.05, sim_sd, niter = 5000,
  nrep = 5000, mc.cores = 6)
}
\arguments{
\item{historical}{The historical (adult) data used for borrowing information under commensurate designs.}

\item{w}{The prespecified expert opinion weight in the payoff function that trades off the competing decisions, between 0 and 1, default 0.5.}

\item{IA_num}{All possible interim analysis timings to choose from, maintain balanced arm at IA.}

\item{N}{The maximum total sample size if the current trial runs to completion.}

\item{p_l}{Early futility rule, that is the required lowest probability at the interim look,
that the novel treatment arm is better than prespecified minimally efficacy \code{delta_min} to declare the trial continue}

\item{p_0}{Final winner rule, that is the required lowest probability at the trial completion,
that the novel treatment arm is better than the control arm to declare the trial succeed}

\item{p_u}{Early winner rule, that is the required lowest probability at the interim look,
that the novel treatment arm is better than the control arm to declare the trial succeed.}

\item{delta_0}{The efficacy under null hypohtesis, default 0.}

\item{delta_1}{The efficacy under targeted alternative hypohtesis.}

\item{delta_design}{The efficacy under a selected design prior.}

\item{delta_min}{The prespecified minimally efficacy to declare the trial continue in early futility rule.}

\item{type1_level}{The targeted level for the calibrated Type I error, default 0.05.}

\item{sim_sd}{The expected standard deviation under a selected design prior.}

\item{niter}{The number of interations for the Bayesian design MCMC steps, default 5000.}

\item{nrep}{The number of repeated calibration for the type I errors and powers calculation, default 5000.}

\item{mc.cores}{The number of computing cores used, default 6}
}
\value{
The function returns a list contains:

\code{result} The optimized timing of interim analysis under the Bayesian adaptive commensurate designs.

\code{data} Additional data table for all calibrated designs under all possible IA timings for further study;
the table contains \code{PP1} and \code{PP2} are the marginal probabilities of early futility under null hypothesis and early efficacy under alternative hypothesis;
\code{cost} and \code{benefit} are the denominator and numerator of payoff function and \code{payoff} is the corresponding value of payoff. We consider two types of payoffs;
\code{var_cost} is the variance of costs;
\code{IA_stop} is the overall probability of stopping at IA;
\code{p1fin} and \code{p2fin} are the calibrated overall type I errors and powers;
\code{ess1} and \code{ess2} are the expected historical sample sizes for two arms respectively;
\code{PP1_Bayes} and \code{PP2_Bayes} are the marginal probabilities of early futility and early efficacy under selected design prior.)
}
\description{
The function to find the optimized timing for the Bayesian adaptive commensurate designs.
}
